+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 ## Unreleased
 
 
+## 2.1.7 (2019-12-15)
+
+### carlton
+
+#### Fixed
+
+ - Added missing scalebox
+ - Re-numbered tail as part 5
+
+
 ## 2.1.6 (2019-11-24)
 
 ### utils