+++ packages/simon/CHANGELOG.md
@@ -6,6 +6,11 @@
 **Note:** Version bump only for package simon
 
 
+## 2.1.7 (2019-12-15)
+
+**Note:** Version bump only for package simon
+
+
 ## 2.1.6 (2019-11-24)
 
 **Note:** Version bump only for package simon